Remove LoadBalancer type from Service
On GKE, services with the LoadBalancer type will automatically create an external IP address
and network load balancer to direct traffic to the Service. Thus, this is exposing the
neo4j cluster to the internet by default.

This change makes the service a normal Service, which works fine for routing within the Kubernetes cluster.